移動方法を少し変えた
