+++ ClientProject/Mof/RachetNClankCopy/Resource/imgui.ini
@@ -1,0 +1,15 @@
+[Window][Debug##Default]
+Pos=60,60
+Size=400,400
+Collapsed=0
+
+[Window][srhygresf]
+Pos=60,60
+Size=37,48
+Collapsed=0
+
+[Window][GameManager]
+Pos=-296,115
+Size=498,207
+Collapsed=0
+
