+++ ClientProject/Mof/RachetNClankCopy/Resource/save/savedata.json
@@ -1,1 +1,1 @@
-{"param":{"money":984},"weapons":["OmniWrench","BombGlove","Pyrocitor"]}
+{"param":{"money":1039},"weapons":["OmniWrench","BombGlove","Pyrocitor"]}
